Finalized layout and submitted to pypi
