+++ vmanage/cli/import_cmd/templates.py
