+++ vmanage/cli/show/device.py
