+++ vmanage/cli/show/template.py
